requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-Bcrypt
GeoAlchemy2
Shapely
GeoPandas